refactor(ingest): extract common/document modules; add Walker trait (#34)
